requirements.txt
selenium
python-oauth2
fastapi
uvicorn
python-jose
pymongo
apscheduler